test(vetting): Wave 4 — intensity/moments/neighbor/imq, incl. first split
Renames: firstorder (test_pixel_intensity_features.h->_regression, test_ibsi_intensity.h
->_ibsi, test_compat_3d_fo_radiomics.h->test_3d_firstorder_pyradiomics.h),
intensity_histogram/neighbor/imq -> *_regression.h. test_3d_inten.h orphaned (flag A).

First genuine SPLIT: test_2d_geometric_moments.h -> test_moments_common.h +
test_moments_skimage.h (oracle_3p) + test_moments_regression.h (unvetted). Byte-for-byte
preserved; common-header avoids ODR dup in the single test_all.cc TU.

Verified in conda build: full gtest suite 696/696 -- NO TESTS DROPPED. Registry
current_test repointed (split-aware); MIGRATION.md 5.11 records it.
