+++ tests/test_all.cc
@@ -4,23 +4,24 @@
 #include "../src/nyx/environment.h"
 #include "../src/nyx/globals.h"
 #include "test_contour.h"
-#include "test_pixel_intensity_features.h"
-#include "test_intensity_histogram.h"
+#include "test_firstorder_regression.h"
+#include "test_intensity_histogram_regression.h"
 #include "test_hu_analytic.h"
 #include "test_hu_mechanics.h"
 #include "test_hounsfield_loader.h"
 #include "test_morphology_features.h"
 #include "test_shape_morphology_2d.h"
-#include "test_2d_geometric_moments.h"
+#include "test_moments_skimage.h"
+#include "test_moments_regression.h"
 #include "test_2d_remaining_features.h"
-#include "test_neighbors_2d.h"
+#include "test_neighbor_regression.h"
 #include "test_initialization.h"
 #include "test_glcm_ibsi.h"
 #include "test_gldm_ibsi.h"
 #include "test_glrlm_ibsi.h"
 #include "test_gldzm_ibsi.h"
 #include "test_glszm_ibsi.h"
-#include "test_ibsi_intensity.h"
+#include "test_firstorder_ibsi.h"
 #include "test_ngldm_ibsi.h"
 #include "test_ngtdm_ibsi.h"
 #include "test_glcm_regression.h"
@@ -31,13 +32,13 @@
 #include "test_ngtdm_regression.h"
 #include "test_roi_blacklist.h"
 #include "test_tiff_loader.h"
-#include "test_image_quality.h"
+#include "test_imq_regression.h"
 #include "test_3d_nifti.h"
 #include "test_omezarr.h"
 #include "test_3d_shape.h"
 #include "test_3d_gldzm_ibsi.h"
 #include "test_3d_ngldm_ibsi.h"
-#include "test_compat_3d_fo_radiomics.h"
+#include "test_3d_firstorder_pyradiomics.h"
 #include "test_3d_glcm_pyradiomics.h"
 #include "test_3d_gldm_pyradiomics.h"
 #include "test_3d_ngtdm_pyradiomics.h"
