+++ tests/test_intensity_histogram_regression.h
